+++ app/Recipe Robot/Filepathable.swift
@@ -1,0 +1,90 @@
+//
+//  Filepathable.swift
+//  Created by Eldon on 11/8/15.
+//  Copyright © 2015 EEAapps.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+
+import Foundation
+
+protocol Filepathable {
+    var path: String { get }
+    var isFile: Bool { get }
+    var isDirectory: Bool { get }
+    var isExecutable: Bool { get }
+}
+
+protocol FilepathValidator: Filepathable {
+    func markAsValidDirectory() -> Bool
+    func markAsValidFile() -> Bool
+    func markAsValidExecutable() -> Bool
+
+}
+
+extension Filepathable {
+    var isFile: Bool {
+        return NSFileManager.defaultManager().fileExistsAtPath(self.path)
+    }
+
+    var isDirectory: Bool {
+        var isDir: ObjCBool = false
+        return NSFileManager.defaultManager().fileExistsAtPath(self.path, isDirectory: &isDir) && isDir
+    }
+
+    var isExecutable: Bool {
+        return NSFileManager.defaultManager().isExecutableFileAtPath(self.path)
+    }
+}
+
+extension String: Filepathable {
+    var path: String { return self }
+}
+
+extension NSTextField: FilepathValidator {
+    var path: String { return self.stringValue }
+
+    private func valid(valid: Bool) -> Bool {
+        textColor = valid ? NSColor.blackColor():NSColor.redColor()
+
+        guard let cell = cell as? NSTextFieldCell,
+            string = cell.placeholderAttributedString?.string ?? cell.placeholderString
+        else {
+            return valid
+        }
+
+        var attrs: [String: AnyObject]?
+        if cell.placeholderAttributedString?.string.characters.count > 0 {
+            attrs = cell.placeholderAttributedString!.attributesAtIndex(0, effectiveRange: nil)
+        } else if cell.attributedStringValue.string.characters.count > 0 {
+            attrs = cell.attributedStringValue.attributesAtIndex(0, effectiveRange: nil)
+        }
+
+        attrs?[NSForegroundColorAttributeName] = valid ? NSColor.lightGrayColor():NSColor.redColor()
+        cell.placeholderAttributedString = NSAttributedString(string: string, attributes: attrs)
+        placeholderAttributedString = cell.placeholderAttributedString
+        
+        return valid
+    }
+
+    func markAsValidDirectory() -> Bool {
+        return valid(path.isDirectory)
+    }
+
+    func markAsValidFile() -> Bool {
+        return valid(path.isFile)
+    }
+
+    func markAsValidExecutable() -> Bool {
+        return valid(path.isExecutable)
+    }
+}